+++ src/handlers/sendCommandTextSanitizer.js
